requirements.txt
langchain==0.0.267
streamlit==1.23.1
groq==0.4.0
pypdf==3.12.2
python-dotenv==1.0.0
faiss-cpu==1.7.4
tiktoken==0.4.0
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
sentence-transformers==2.2.2
arxiv==1.4.7
PyPDF2==3.0.1
nltk==3.8.1
requests==2.31.0
urllib3==2.0.7
